(feat) Call addRatingsToSearchResults after retrieving factual places, send back all results
