+++ server/api/mongo/placeModel.js
@@ -1,5 +1,7 @@
 var Q = require('q');
 var mongodb = require('./../../mongoConfig.js');
+var _ = require('lodash');
+var User = require('./userModel');
 var Place = {};
 
 Place.createOrUpdatePlace = function(placeData) {
@@ -53,6 +55,58 @@ Place.setPropertyOnPlaceDocument = function(factual_id, key, value) {
       deferred.resolve(result);
     }
   });
+};
+
+Place.addRatingsToSearchResults = function (factualPlaces, ratedPlaces) {
+  // Getting rater's info after retrieving factual results is more efficient
+  // than beforehand (in getRatedPlaces) as you may not need to 
+  // iterate through all rated places, but only rated places matching the
+  // search params. 
+
+  // Store ratings, avgRating, and facebook ids for each place (with factual ID as the key)
+  // Retrieve names for all facebook ids in search results
+  // Store names in ratingInfoByFactualID
+  // Use ratingInfoByFactualID when rendering search results (lookup on factual ID)
+  var deferred = Q.defer();
+  var allFacebookIDs = [];
+  var ratingInfoByFactualID = {};
+  var raters = {};
+
+  // Iterate through factual results
+  // Organize of users who have rated each place
+  _.each(factualPlaces, function(place) {
+    var ratings = ratedPlaces[place.factual_id].ratings;
+    var facebookIDs;
+
+    if (ratings.length === 1) {
+      facebookIDs = [ratings[0].facebookID];
+      
+    } else if (ratings.length > 1) {
+      facebookIDs = [ratings[0].facebookID, ratings[1].facebookID];
+    }
+
+    ratingInfoByFactualID[place.factual_id] = {
+      facebookIDs: facebookIDs,
+      ratings: ratings,
+      avgRating: ratedPlaces[place.factual_id].avgRating
+    };
+    allFacebookIDs = allFacebookIDs.concat(facebookIDs);
+  });
+  
+  allFacebookIDs = _.uniq(allFacebookIDs);
+
+  User.getNames(allFacebookIDs)
+    .then(function(users) {
+      _.each(users, function (user) {
+        raters[user.facebookID] = user.name;
+      });
+
+      deferred.resolve({ factualPlaces: factualPlaces, raters: raters, ratingInfoByFactualID: ratingInfoByFactualID});
+    })
+    .catch(function(err) {
+      console.log(err);
+    });
+  return deferred.promise;
 };
 
 module.exports = Place;
