+++ server/api/neo4j/placeModel.js
@@ -441,58 +441,4 @@ Place.discoverByCategoryOrNameAndLocation = function (facebookID, locationTerm, 
   return deferred.promise;
 };
 
-Place.addRatingsToSearchResults = function (factualPlaces, ratedPlaces) {
-  // Getting rater's info after retrieving factual results is more efficient
-  // than beforehand (in getRatedPlaces) as you may not need to 
-  // iterate through all rated places, but only rated places matching the
-  // search params. 
-
-  // Store ratings, avgRating, and facebook ids for each place (with factual ID as the key)
-  // Retrieve names for all facebook ids in search results
-  // Store names in ratingInfoByFactualID
-  // Use ratingInfoByFactualID when rendering search results (lookup on factual ID)
-  var deferred = Q.defer();
-  var allFacebookIDs = [];
-  var ratingInfoByFactualID = {};
-  var raters = {};
-
-  // Iterate through factual results
-  // Organize of users who have rated each place
-  _.each(factualPlaces, function(place) {
-    var ratings = ratedPlaces[place.factualID].ratings;
-    var facebookIDs;
-
-    if (ratings.length === 1) {
-      facebookIDs = [ratings[0].facebookID];
-      
-    } else if (ratings.length > 1) {
-      facebookIDs = [ratings[0].facebookID, ratings[1].facebookID];
-    }
-
-    ratingInfoByFactualID[place.factualID] = {
-      facebookIDs: facebookIDs,
-      ratings: ratings,
-      avgRating: ratedPlaces[place.factualID.avgRating]
-    };
-    // ratingInfoByFactualID[place.factualID].facebookIDs = facebookIDs;
-    // place.ratings = ratings;
-    // place.avgRating = ratedPlaces[place.factualID.avgRating];
-    allFacebookIDs = allFacebookIDs.concat(facebookIDs);
-
-  });
-
-  User.getNames(allFacebookIDs)
-    .then(function(persons) {
-      _.each(persons, function (person) {
-        raters[person.factualID] = person.name;
-      });
-
-      deferred.resolve({ raters: raters, ratingInfoByFactualID: ratingInfoByFactualID});
-    })
-    .catch(function(err) {
-      console.log(err);
-    });
-  return deferred.promise;
-};
-
 module.exports = Place;
